implement popup dialog

+++ src/app/home/home.component.ts
@@ -1,8 +1,10 @@
+import { ConfirmDeleteDialogComponent } from './../confirm-delete-dialog/confirm-delete-dialog.component';
 import { Component, OnInit } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
+import { MatDialog, MatDialogRef } from '@angular/material/dialog';
 import { MatSelectChange } from '@angular/material/select';
 import { Todo } from '../models/todo.model';
 import { ApiService } from '../services/api.service';
+import { TodoComponent } from '../todo/todo.component';
 
 @Component({
   selector: 'app-home',
@@ -20,6 +22,27 @@ export class HomeComponent implements OnInit {
   ngOnInit(): void {
     // get all todos from api in the beginning
     this.getAllTodosFromAPI();
+  }
+
+  openPopup() {
+    const dialogRef: MatDialogRef<TodoComponent> = this.dialog.open(
+      TodoComponent,
+      {
+        width: '500px',
+        hasBackdrop: true,
+        role: 'dialog',
+      }
+    );
+
+    dialogRef.afterClosed().subscribe((data) => {
+      this.apiService
+        .createTodo(data.title, data.description)
+        .subscribe((res: any) => {
+          console.log(res);
+          this.todos.push(res);
+          this.filteredTodos = this.todos;
+        });
+    });
   }
 
   // get all todos from api
@@ -53,11 +76,18 @@ export class HomeComponent implements OnInit {
 
   deleteTodo(e: Event, todoId: number) {
     e.stopPropagation();
-    this.apiService.deleteTodo(todoId).subscribe((res) => {
-      // @ts-ignore
-      if (res.success) {
-        this.todos = this.todos.filter((todo) => todo.id !== todoId);
-        this.filteredTodos = this.todos;
+
+    const dialogRef = this.dialog.open(ConfirmDeleteDialogComponent);
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.apiService.deleteTodo(todoId).subscribe((res) => {
+          // @ts-ignore
+          if (res.success) {
+            this.todos = this.todos.filter((todo) => todo.id !== todoId);
+            this.filteredTodos = this.todos;
+          }
+        });
       }
     });
   }
